pkg/crypto: remove incorrect test assertion
It is legal to decrypt a blob of data with a different key
than what was used to initial encrypt it. The operation
can succeed, but the decrypted data will simply not match
the original plaintext.

This assertion was actually catching a padding error, which
was not always getting triggered due to the call to
crypto.Rand in AESEncrypt. It is safe to simply compare the
decrypted data and ensure it does not match the initial
plaintext.