db: uniqueness constraint for email
The default transaction isolation mode in Postgres is "Read Committed",
where nonrepeatable reads are possible. So that means in this case two
different transactions could be trying to create a user with the same
email address, and both could `commit` successfully, even though a
`select` at the beginning of the txn checked for users with that email
address.

http://www.postgresql.org/docs/9.1/static/transaction-iso.html